rename(ui-design-tool): rename SelectionOverlay to HoveringOverlay
