+++ packages/ui-design-tool/components/HoveringOverlay/HoveringOverlay.css.ts
